Update : #108 TodoList에서 DnD 처리 함수 파일 분할
- 기존에 UI 코드 파일에 DnD 처리 코드가 많은 공간을 잡아먹고 있어서 유지 보수를 하기에 불편함이 있었습니다.
- 이를 해결하기 위해 TodoList 파일을 만들고 해당 파일 안에 UI를 담당하는 index파일, dnd 핸들링을 담당하는 dragHelper파일을 만들었습니다.
- mock 데이터도 따로 파일을 만들었습니다.
